feat(pages): Parametrize Ingredients page
